reafactor: Container props에 대한 정의를 types/로 이동
